Reworked REST endpoints
