+++ Vapour.Server.Interfaces/Controller/ControllerHostStatusResponse.cs
@@ -1,0 +1,6 @@
+﻿namespace Vapour.Server.Controller;
+
+public class ControllerHostStatusResponse
+{
+    public bool IsRunning { get; init; }
+}
